fix: adds update schedule endpoint

+++ footballseason/management/commands/generate_schedule.py
@@ -21,9 +21,12 @@ LOG = logging.getLogger(__name__)
 # Call from CLI via: $ python manage.py generate_schedule
 class Command(BaseCommand):
     season = fb_utils.get_season()
-    week_list = range(1, fb_utils.NUM_WEEKS + 1)
+    current_week = fb_utils.get_week()
+    week_list = range(current_week, fb_utils.NUM_WEEKS + 1)
 
-    def add_games_from_one_week(self, season, week):
+    def add_games_from_one_week(self, season, week, update_only, dry_run):
+        updated_games = 0
+
         url = "http://espn.go.com/nfl/schedule/_/year/{0}/week/{1}".format(season, week)
         LOG.info("url: {0}".format(url))
         with urllib.request.urlopen(url) as response:
@@ -81,27 +84,63 @@ class Command(BaseCommand):
                     except Game.DoesNotExist:
                         obj = Game(season=season, week=week, home_team=home, away_team=away, game_time=game_datetime)
                         LOG.info("Adding: {0}".format(obj))
+                        updated_games += 1
                     else:
                         if obj.game_time != game_datetime:
                             obj.game_time = game_datetime
                             LOG.info(f"{obj} was already on the schedule, updating gametime and saving")
+                            updated_games += 1
                     finally:
-                        obj.save()
+                        if not dry_run:
+                            obj.save()
                         all_game_ids.append(obj.id)
+
+        if len(all_game_ids) < 10:
+            # There either was a problem finding the games or this week already happened
+            # Either way, skip removing the games
+            LOG.info("Skipping week with not enough matched game times")
+            return 0
 
         removed_games = Game.objects.filter(season=season, week=week).exclude(id__in=all_game_ids)
 
         if removed_games:
+            updated_games = removed_games.count()
             try:
                 LOG.info(f"Removing {removed_games.count()} games: {list(removed_games)}")
-                removed_games.delete()
+                if not dry_run and not update_only:
+                    removed_games.delete()
             except ProtectedError:
                 LOG.info("Unable to remove games with picks")
 
-    def add_all_games(self):
+        return updated_games
+
+    def add_all_games(self, update_only=False, dry_run=False):
+        updated_games = 0
         for week in self.week_list:
             LOG.info("Processing week {0}".format(week))
-            self.add_games_from_one_week(self.season, week)
+            updated_games += self.add_games_from_one_week(self.season, week, update_only, dry_run)
+        return updated_games
+
+    def add_arguments(self, parser):
+        # Named (optional) arguments
+        parser.add_argument(
+            "--update-only",
+            action="store_true",
+            help="Only allow updates, no deletions",
+        )
+        parser.add_argument(
+            "--dry-run",
+            action="store_true",
+            help="Just print out what would happen, do not make any changes",
+        )
 
     def handle(self, *args, **options):
-        self.add_all_games()
+        update_only = options.get("update_only")
+        dry_run = options.get("dry_run")
+        if dry_run:
+            LOG.info("dry-run option is enabled")
+        if update_only:
+            LOG.info("update-only option is enabled")
+
+        games_updated = self.add_all_games(update_only=update_only, dry_run=dry_run)
+        return f"Updated {games_updated} games"
